Adding validate Comment and stock Controller

+++ api/Controllers/CommentController.cs
@@ -24,13 +24,15 @@ namespace api.Controllers
         [HttpGet]
         public async Task<IActionResult> GetAll()
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var comment = await _commentRepo.GetAllAsync();
             var commentDto = comment.Select(s => s.ToCommentDto());
             return Ok(commentDto);
         }
-        [HttpGet("{id}")]
+        [HttpGet("{id:int}")]
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var comment = await _commentRepo.GetByIdAsync(id);
             if (comment == null)
             {
@@ -39,25 +41,28 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
         [HttpDelete]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var commentModel = await _commentRepo.DeleteAsync(id);
             if (commentModel == null) return NotFound("Comment not found!!!");
             return NoContent();
         }
-        [HttpPost("{stockId}")]
+        [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             if (!await _stockRepo.ExistStock(stockId)) return BadRequest("Stock does not exist!!!!");
             var commentModel = commentDto.ToCreateCommentDto(stockId);
             await _commentRepo.CreateAsync(commentModel);
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
         [HttpPut]
-        [Route("{id}")]
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentDto updateDto)
         {
+            if (!ModelState.IsValid) return BadRequest(ModelState);
             var commentModel = await _commentRepo.UpdateAsync(id, updateDto);
             if (commentModel == null) return NotFound("Comment not found!!!");
